minetest: Allow building just the client or server and allow building without sound.
Building only the server requires irrlicht source available, but doesn't require it built, which would also require X installed.  I set this spell to download irrlicht if needed vs relying on the other to have a source-only option.  The source is always included in DETAILS because it will be needed regardless, but it's only extracted in PRE_BUILD if it's needed.